requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mypy==0.610
numpy==1.14.5
pandas==0.23.3
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.9
typed-ast==1.1.0
Werkzeug==0.14.1
pytest==3.6.3
pytest-cov==2.5.1
